+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Added:
 - Settings to condense multiple consecutive server messages into a single abbreviated message
 - `/detach` command to detach from channels (soju feature)
 - Support for USERINFO CTCP query
+- Display own nickname next to text input field
 
 Fixed: